fixing many

+++ app/Http/Controllers/ProfileController.php
@@ -29,7 +29,9 @@ class ProfileController extends Controller
             $name = $file->getClientOriginalName();
             $file_name = date('mdYHis') . '-' . $name;
             $file = $file->storeAs('file', $file_name, 'public_uploads');
-            $user->deleteFile();
+            if (Auth::user()->image) {
+                $user->deleteFile();
+            }
             $data['image'] = $file;
         };
         if ($request->password) {
